Factor out webpack plugin builder
